src/data/make_processed.py
# -*- coding: utf-8 -*-
import os
import logging
from pathlib import Path
from dotenv import find_dotenv, load_dotenv


def makeProcessed():
    """ 
        Runs data processing scripts to turn raw data from (/projectdir/data/raw) into
        cleaned data ready to be analyzed (saved in /projectdir/data/processed).
    """
    logger = logging.getLogger(__name__)
    logger.info('making final data set from raw data')
    
    input_filepath = Path("../../data/raw")


if __name__ == '__main__':
    log_fmt = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    logging.basicConfig(level=logging.INFO, format=log_fmt)

    # find .env automagically by walking up directories until it's found, then
    # load up the .env entries as environment variables
    load_dotenv(find_dotenv())

    makeProcessed()
