requirements.txt
# local package
-e .

# external requirements
kaggle
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
numpy==1.19.5
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.5.0
torch==1.9.0
matplotlib==3.4.2
seaborn==0.11.1